pyproject.toml
[tool.poetry]
name = "rag-virtual-assistant-course"
version = "0.1.0"
description = "A tutorial series on building RAG-based virtual assistants."
authors = ["thanos pantos <pantos.thn@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
langchain = "^0.3.24"
langchain-community = "^0.3.23"
langchain-openai = "^0.3.14"
langchainhub = "^0.1.21"
chromadb = "^1.0.7"
tiktoken = "^0.9.0"
notebook = "^7.4.1"
python-dotenv = "^1.1.0"
sentence-transformers = "^4.1.0"
langchain-huggingface = "^0.1.2"
ipywidgets = "^8.1.6"
langgraph = "^0.4.1"
requests = "^2.32.3"
gradio = "^5.28.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
